LayoutTests/fast/css/border-shorthand-initialize-longhands.html

<!DOCTYPE html>
<html>
<body>
<script src="../js/resources/js-test-pre.js"></script>
<script>

description("This test ensures border shorthand property initializes longhand properties such as border-top-style and border-right-color.");

var div = document.createElement('div');
shouldBe("div.setAttribute('style', 'border: 1px');div.style.borderTopStyle", "'initial'");
shouldBe("div.style.borderTopStyle", "'initial'");
shouldBe("div.style.borderRightStyle", "'initial'");
shouldBe("div.style.borderBottomStyle", "'initial'");
shouldBe("div.style.borderLeftStyle", "'initial'");

shouldBe("div.style.borderTopColor", "'initial'");
shouldBe("div.style.borderRightColor", "'initial'");
shouldBe("div.style.borderBottomColor", "'initial'");
shouldBe("div.style.borderLeftColor", "'initial'");

shouldBe("div.setAttribute('style', 'border: solid');div.style.borderTopWidth", "'initial'");
shouldBe("div.style.borderRightWidth", "'initial'");
shouldBe("div.style.borderBottomWidth", "'initial'");
shouldBe("div.style.borderLeftWidth", "'initial'");

</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
